requirements.txt
chardet==5.1.0
cryptography==38.0.4
cryptography.egg==info
fuzzywuzzy==0.18.0
nltk==3.8
pdfplumber==0.11.4
PyMySQL==1.1.1
python-dotenv==1.0.1
python-telegram-bot==21.6
python_docx==1.1.2
spacy==3.7.6
